==> src/1.DataPrepration.ipynb <==
from google.colab import drive
drive.mount('/content/drive/')

import csv
import os
import random
from ast import main
import pandas as pd

mainData=[]

with open("/content/drive/MyDrive/Sentiment-Analysis/Data/main/Dataset-SA-w.o null.csv", 'r') as csv_file:
    # Create a CSV reader object
    csv_reader = csv.reader(csv_file)
    # Read and print the header row
    header = next(csv_reader)
    # Loop through each data row in the CSV file
    for row in csv_reader:
      mergedReview= row[3]+' '+row[4]
      mainData.append([mergedReview,row[5]])


random.shuffle(mainData)

with open('MainData.csv', 'w', newline='') as csv_file:
    csv_writer = csv.writer(csv_file)
    for row in mainData:
        csv_writer.writerow(row)

dataset = pd.read_csv("MainData.csv")
dataset.columns = ['review', 'sentiment']


dataset.to_csv("/content/drive/MyDrive/Sentiment-Analysis/Data/main/MainData.csv")
os.remove("/content/MainData.csv")

len(dataset)

train, dev, test=[],[],[]

for ind in range(164032):
  train.append(mainData[ind])

for ind in range(164032,184536):
  dev.append(mainData[ind])

for ind in range(184536,205040):
  test.append(mainData[ind])

with open('/content/drive/MyDrive/Sentiment-Analysis/Data/w. Neutral Splits/TrainFile.csv', 'w', newline='') as csv_file:
    csv_writer = csv.writer(csv_file)
    for row in test:
      csv_writer.writerow(row)


with open('/content/drive/MyDrive/Sentiment-Analysis/Data/w. Neutral Splits/TrainFile.csv', 'w', newline='') as csv_file:
    csv_writer = csv.writer(csv_file)
    for row in train:
      csv_writer.writerow(row)


with open('/content/drive/MyDrive/Sentiment-Analysis/Data/w. Neutral Splits/DevFile.csv', 'w', newline='') as csv_file:
    csv_writer = csv.writer(csv_file)
    for row in dev:
      csv_writer.writerow(row)

with open('/content/drive/MyDrive/Sentiment-Analysis/Data/wo. Neutral Splits/TestFile.csv', 'w', newline='') as csv_file:
    csv_writer = csv.writer(csv_file)
    for row in test:
      if row[1] != 'neutral':
        csv_writer.writerow(row)


with open('/content/drive/MyDrive/Sentiment-Analysis/Data/wo. Neutral Splits/TrainFile.csv', 'w', newline='') as csv_file:
    csv_writer = csv.writer(csv_file)
    for row in train:
      if row[1] != 'neutral':
        csv_writer.writerow(row)


with open('/content/drive/MyDrive/Sentiment-Analysis/Data/wo. Neutral Splits/DevFile.csv', 'w', newline='') as csv_file:
    csv_writer = csv.writer(csv_file)
    for row in dev:
      if row[1] != 'neutral':
        csv_writer.writerow(row)
