==> src/2.Lexicon-Based.ipynb <==
!pip install vaderSentiment

from google.colab import drive
drive.mount('/content/drive')

import time
import os
import json
from sklearn.metrics import accuracy_score , precision_score, recall_score , f1_score
import pandas as pd
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

test_data = pd.read_csv("/content/drive/MyDrive/Sentiment-Analysis/Data/wo. Neutral Splits/TestFile.csv", header=None)
test_data.columns = ['review', 'sentiment']

# test_data = test_data.drop(0)

analyzer = SentimentIntensityAnalyzer()

def calculate_sentiment_scores(sentence):
    sntmnt = analyzer.polarity_scores(sentence)['compound']
    return(sntmnt)

start = time.time()

eng_snt_score =  []

for review in test_data['review']:
    snts_score = calculate_sentiment_scores(review)
    eng_snt_score.append(snts_score)

end = time.time()

test_data['sentiment_score'] = eng_snt_score
vader_sentiment = [ ]
for i in range(len(test_data)):
    sentiment_score = test_data.iloc[i]['sentiment_score']

    if sentiment_score > 0:
        vader_sentiment.append('positive')
    elif sentiment_score <= 0:
        vader_sentiment.append('negative')

test_data['vader_sentiment_labels'] = vader_sentiment


test_data['actual_label'] = test_data.iloc[:, 1].map({'positive': 0, 'negative':1})
test_data['predicted_label'] = test_data['vader_sentiment_labels'].map({'positive': 0, 'negative':1})

with open("/content/drive/MyDrive/Sentiment-Analysis/Data/evaluation.json", 'r') as json_file:
          df = json.loads(json_file.read())

for i, key in enumerate(df.keys()):
        df[key]['lexicon'] = int(test_data['predicted_label'][i])
        df[key]['actual'] = int(test_data['actual_label'][i])

with open("/content/drive/MyDrive/Sentiment-Analysis/Data/evaluation.json", "w") as json_file:
    json.dump(df, json_file)

y_act = test_data['actual_label'].values
y_pred = test_data['predicted_label'].values
accuracy = accuracy_score(y_act, y_pred)
print(f'acc:{accuracy}')
precision = precision_score(y_act, y_pred, average='macro')
print(f'precision: {precision}')
recall = recall_score(y_act, y_pred, average='macro')
print(f'recall:{recall}')
f1 = f1_score(y_act, y_pred, average='macro')
print(f'f1:{f1}')



precision_w = precision_score(y_act, y_pred, average='weighted')
print(f'precision: {precision_w}')
recall_w = recall_score(y_act, y_pred, average='weighted')
print(f'recall:{recall_w}')
f1_w = f1_score(y_act, y_pred, average='weighted')
print(f1_w)
