==> src/8.Evaluation.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import json
import pandas as pd
import spacy

from collections import Counter
from sklearn.metrics import precision_score, recall_score, f1_score


content = open('/content/drive/MyDrive/Sentiment-Analysis/Data/evaluation.json')
data = json.load(content)


majority = []
for value in data.values():
  majority.append(max(Counter(value.values()), key=lambda key: Counter(value.values())[key]))


for i, key in enumerate(data.keys()):
  data[key]['Majority'] = majority[i]

actual_labels = []
majority_labels = []
for value in data.values():
  actual_labels.append(value['actual'])
  majority_labels.append(value['Majority'])

f1 = f1_score(actual_labels, majority_labels)
recall = recall_score(actual_labels, majority_labels)
precision = precision_score(actual_labels, majority_labels)

f1,precision, recall

with open("/content/drive/MyDrive/Sentiment-Analysis/Data/evaluation.json", "w") as json_file:
    json.dump(data, json_file)


#longest reviews in dataset
df = pd.read_csv("/content/drive/MyDrive/Sentiment-Analysis/Data/wo. Neutral Splits/TestFile.csv")
df.columns = ['review', 'sentiment']

nlp = spacy.load('en_core_web_sm')

top_25 ={}
for index, row in df.iterrows():
  if len(nlp(row['review'])) > 75:
     top_25[row['review']] = row["sentiment"]

# There is something wrong with first row, so drop it for now and then you have to fix it
df1 = pd.DataFrame.from_dict(top_25, orient='index', columns=['sentiment']).reset_index()
df1.drop(index=0)


df1.to_csv("/content/drive/MyDrive/Sentiment-Analysis/Data/wo. Neutral Splits/long-review-test.csv", index=False)
